Fix softmax

+++ src/frontends/pytorch/src/op_table.cpp
@@ -309,7 +309,7 @@ const std::map<std::string, CreatorFunction> get_supported_ops() {
          [](NodeContext& context) -> OutputVector {
              auto axis = context.const_input<int64_t>(1);
              return {
-                 context.mark_node(std::make_shared<opset8::Softmax>(context.get_input(0), static_cast<size_t>(axis)))};
+                 context.mark_node(std::make_shared<opset8::Softmax>(context.get_input(0), axis))};
          }},
 
         //{"aten::cat", done as transformation},
